requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
motor==3.3.1
python-dotenv==1.0.0
